further work on tests; #55
