+++ tests/test_config_unbound_env_vars.py
@@ -9,9 +9,9 @@ __author__ = "Vince Reuter"
 __email__ = "vreuter@virginia.edu"
 
 
-@pytest.mark.parametrize(["genome", "asset"], REQUESTS)
+@pytest.mark.parametrize(["genome", "asset", "tag"], REQUESTS)
 @pytest.mark.parametrize("evs", [["NOT_A_VAR"], ["NOT_A_VAR", "RANDNAME"]])
-def test_missing_env_vars_in_genome_config_path_raises_exception(rgc, tmpdir, evs, genome, asset, gencfg,
+def test_missing_env_vars_in_genome_config_path_raises_exception(rgc, tmpdir, evs, genome, asset, tag, gencfg,
                                                                  remove_genome_folder, cfg_file_copy):
     """ Unbound env var(s) in genome folder path cause error. """
     assert all(_is_unbound(v) for v in evs)
@@ -22,7 +22,7 @@ def test_missing_env_vars_in_genome_config_path_raises_exception(rgc, tmpdir, ev
     assert path == rgc[CFG_FOLDER_KEY]
     assert not os.path.exists(path)
     with pytest.raises(UEVErr) as err_ctx:
-        rgc.pull_asset(genome=genome, asset=asset, tag=None, genome_config=cfg_file_copy, get_json_url=get_get_url(genome, asset))
+        rgc.pull_asset(genome=genome, asset=asset, tag=tag, genome_config=cfg_file_copy, get_json_url=get_get_url(genome, asset))
     err_msg = str(err_ctx.value)
     print("Observed error message: {}".format(err_msg))
     missing = [v for v in evs if v not in err_msg]
